Make it easier to produce differently-branded versions of CoreOS
If anybody wants to derive from CoreOS, it's helpful for us to make it as
easy as possible for them to strip the branding so there's less risk of
user confusion. As a first step, move the branding used in the build process
to a separate file and ensure that images end up named appropriately.